Work around for https://github.com/JuliaLang/julia/issues/19979
Retry.jl used to use try/catch inside the catch block's "@ignore if ..."
condition. The purpose of this was to treat if conditions that throw
errors as false. e.f. @ignore if e.reason == "Foo" end might throw an
exception if e does not have a reason field. Retry.jl used to allow this
and treat it as false.

At present doing a nested try/catch in a catch block causes the Julia
stacktrace to be corrupted: https://github.com/JuliaLang/julia/issues/19979#issuecomment-344439163

After this commit, Retry.jl not requires @ignore if conditions to return
true or false and not throw exceptions.

The ecode() function has been exported to allow replacing

    @ignore if e.code == "404" end

with

    @ignore if ecode(e) == "404" end